ARIES-1453 extract in separate test class
