+++ subsystem/subsystem-core/src/test/java/org/apache/aries/subsystem/core/archive/FragmentHostHeaderTest.java
@@ -21,24 +21,9 @@ package org.apache.aries.subsystem.core.archive;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.fail;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Set;
-
-import org.apache.aries.subsystem.core.capabilityset.CapabilitySet;
-import org.apache.aries.subsystem.core.capabilityset.SimpleFilter;
-import org.junit.Assert;
 import org.junit.Test;
 import org.osgi.framework.Version;
 import org.osgi.framework.VersionRange;
-import org.osgi.framework.namespace.BundleNamespace;
-import org.osgi.framework.namespace.HostNamespace;
-import org.osgi.framework.namespace.PackageNamespace;
-import org.osgi.resource.Capability;
-import org.osgi.resource.Namespace;
-import org.osgi.resource.Requirement;
-import org.osgi.resource.Resource;
 
 public class FragmentHostHeaderTest {
 	@Test
@@ -112,51 +97,6 @@ public class FragmentHostHeaderTest {
                 header.getClauses().iterator().next(), 
                 new VersionRange(VersionRange.LEFT_CLOSED, new Version("2.0"), new Version("3.0"), VersionRange.RIGHT_OPEN));
     }
-	
-	@Test
-	public void shouldResolveFragmentHostWithBundleVersionToVersionRangeAttributeAndMatchWithBundlesInThatRange() {
-		FragmentHostHeader header = new FragmentHostHeader("Fragment-Host: host-bundle;bundle-version=\"[9.6.0,10)\"");
-		assertClauses(header, 1);
-        assertSymbolicName(header.getClauses().iterator().next(), "host-bundle");
-        assertBundleVersionAttribute(
-                header.getClauses().iterator().next(), 
-                new VersionRange(VersionRange.LEFT_CLOSED, new Version("9.6.0"), new Version("10"), VersionRange.RIGHT_OPEN));
-		FragmentHostRequirement requirement = new FragmentHostRequirement(header.getClauses().iterator().next(), null);
-		Resource dummyResource = new Resource() {
-			
-			@Override
-			public List<Requirement> getRequirements(String namespace) {
-				return new ArrayList();
-			}
-			
-			@Override
-			public List<Capability> getCapabilities(String namespace) {
-				return new ArrayList();
-			}
-		};
-		FragmentHostCapability capability = new FragmentHostCapability(new BundleSymbolicNameHeader("host-bundle"), new BundleVersionHeader("9.6.1"), dummyResource);
-		dummyResource.getCapabilities("").add(capability);
-		
-		
-		
-		String filterDirective = requirement.getDirectives().get(Namespace.REQUIREMENT_FILTER_DIRECTIVE);
-        SimpleFilter simpleFilter;
-        if (filterDirective == null) {
-            simpleFilter = new SimpleFilter(null, null, SimpleFilter.MATCH_ALL);
-        }
-        else {
-            simpleFilter = SimpleFilter.parse(filterDirective);
-        }
-        String namespace = requirement.getNamespace();
-        CapabilitySet capabilitySet = new CapabilitySet(Arrays.asList(HostNamespace.HOST_NAMESPACE), true);
-        capabilitySet.addCapability(capability);
-        Set<Capability> capabilities = capabilitySet.match(
-                simpleFilter, 
-                PackageNamespace.PACKAGE_NAMESPACE.equals(namespace)
-                        || BundleNamespace.BUNDLE_NAMESPACE.equals(namespace)
-                        || HostNamespace.HOST_NAMESPACE.equals(namespace));
-        Assert.assertTrue(capabilities.size() == 1);
-	}
 	
 	private void assertBundleVersionAttribute(FragmentHostHeader.Clause clause, VersionRange value) {
 	    assertEquals("Wrong bundle version", value, ((BundleVersionAttribute)clause.getAttribute(BundleVersionAttribute.NAME)).getVersionRange());
